Fixes #61 by removing delete icons for users without permission

+++ app/helpers/keywording_helper.rb
@@ -2,8 +2,9 @@ module KeywordingHelper
 
   def manage_keywording(keywording, on_parent = false)
     link = link_to keywording.keyword.name, keywording.keyword
-    link += ' '
-    unless on_parent
+
+    if ! on_parent && policy(keywording).destroy?
+      link += ' '
       link += link_to_delete (fa_icon 'remove'), keywording, method: :delete
     end
     content_tag(:span, bootstrap_label_with_tooltip(link, type: :primary, title: keywording.keyword.description, id: "keywording-#{keywording.id}"))
